+++ webapp/config.py
@@ -1,3 +1,5 @@
+from datetime import timedelta
+
 import os
 
 basedir = os.path.abspath(os.path.dirname(__file__))
@@ -5,3 +7,5 @@ basedir = os.path.abspath(os.path.dirname(__file__))
 SQLALCHEMY_DATABASE_URI = 'sqlite:///' + os.path.join(basedir, '..', 'db.db')
 
 SECRET_KEY = "lisedgtbhrunhgiqeru7345y923n#$%^&*(djfkbg)"
+
+REMEMBER_COOKIE_DURATION = timedelta(days=5)
