+++ webapp/templates/login.html
@@ -24,7 +24,7 @@
                 </div>
                 {% endif %}
               {% endwith %}
-                <form action="{{ url_for('process_login') }}" method="post">
+                <form action="{{ url_for('user.process_login') }}" method="post">
                     {{ form.hidden_tag() }}
                     <div class=form-group>
                         {{ form.username.label }}<br>
@@ -33,6 +33,10 @@
                     <div class=form-group>
                         {{ form.password.label }}<br>
                         {{ form.password() }}
+                    </div>
+                    <div class="form-group form-check">
+                        {{ form.remember_me() }}
+                        {{ form.remember_me.label(class_='form-check-label') }}
                     </div>
                     <p>{{ form.submit }}</p>
                 </form>
