+++ webapp/user/views.py
@@ -1,0 +1,36 @@
+from flask import Blueprint, render_template, flash, redirect, url_for
+from flask_login import current_user, login_user, logout_user
+
+from webapp.user.forms import LoginForm
+from webapp.user.models import User
+
+blueprint = Blueprint('user', __name__, url_prefix='/users')
+
+@blueprint.route('/')
+def login():
+    if current_user.is_authenticated:
+        return redirect(url_for('index'))
+    title = 'Авторизация'
+    login_form = LoginForm()
+    return render_template('login.html', page_title=title, form=login_form)
+
+@blueprint.route('/process-login', methods=['POST'])
+def process_login():
+    form = LoginForm()
+    if form.validate_on_submit():
+        user = User.query.filter_by(login=form.username.data).first()
+        if user and user.check_password(form.password.data):
+            login_user(user, remember=form.remember_me.data)
+            # flash('Вы вошли на сайт')
+            if user.is_admin:
+                return redirect(url_for('admin'))
+            else:
+                return redirect(url_for('index'))
+
+    flash('Неправильное имя пользователя или пароль')
+    return redirect(url_for('user.login'))
+
+@blueprint.route('/logout')
+def logout():
+    logout_user()
+    return redirect(url_for('user.login'))
